+++ pdb2net/__init__.py
@@ -1,3 +1,3 @@
 """PDB2Net package metadata."""
 
-__version__ = "0.2.0rc3"
+__version__ = "0.2.0rc4"